Some progress - yet to figure out a way to store multiple topping inputs into the ArrayList and then print them out as well.
Need to add a loop for the menu viewing to allow viewing of different parts if needs be, currently you can only select one and then it moves on.

Also need to add an extra loop for the toppings' selection to allow customers to confirm they want more toppings or move on to the next stage of ordering if they have enough.